+++ tests/FSharp.Analyzers.Tests/data/partialapp/Warnings are emitted for do exprs.fs
@@ -1,8 +1,0 @@
-module M
-    module S =
-        let myFunc x y = x + y
-
-        do
-            let f = myFunc 1 // should warn
-            myFunc 1 // should warn
-            ()