+++ tests/FSharp.Analyzers.Tests/data/partialapp/Warnings are emitted for module functions.fs
@@ -1,3 +1,0 @@
-module M
-
-let partapp5: (int seq -> int seq) = Seq.map (fun x -> x + 1) // should warn