+++ tests/FSharp.Analyzers.Tests/data/partialapp/Warnings are emitted in simple bindings.fs
@@ -1,5 +1,0 @@
-module M
-
-let myFunc x y = x + y
-
-let partapp2 = myFunc 4 // should warn