+++ tests/FSharp.Analyzers.Tests/data/partialapp/negative/No warning for left piped2 args.fs
@@ -1,5 +1,0 @@
-module M
-
-let myFunc a b c = a + b + c
-
-myFunc 10 <|| (20, 30)