+++ tests/FSharp.Analyzers.Tests/data/partialapp/negative/No warning for right piped args.fs
@@ -1,5 +1,0 @@
-module M
-
-let myFunc a b = a + b
-
-20 |> myFunc 10