+++ tests/FSharp.Analyzers.Tests/data/partialapp/negative/No warning for right piped args2.fs
@@ -1,6 +1,0 @@
-module M
-
-let f x = 3 + x
-let myFunc a b c d = a + b + c + d
-
-30 |> myFunc 10 20 21