+++ tests/FSharp.Analyzers.Tests/data/partialapp/negative/No warning for right piped2 args.fs
@@ -1,5 +1,0 @@
-module M
-
-let myFunc a b c = a + b + c
-
-(20, 30) ||> myFunc 10