+++ tests/FSharp.Analyzers.Tests/data/loggingtemplatemissingvalues/negative/No warnings for unusual template names.fs
@@ -1,0 +1,9 @@
+module M
+
+    open Microsoft.Extensions.Logging
+
+    let testlog () =
+        use factory = LoggerFactory.Create(fun b -> b.AddConsole() |> ignore)
+        let logger: ILogger = factory.CreateLogger("Program")
+
+        logger.Log(LogLevel.Information, "xxx {o-ne} yyy {t_wo}", 23, 42)